generate: Implement wildcard enable/disable commands
If the last character in the instruction name for the enable or
disable command is an asterisk '*', then the command will enable
or disable all instructions that begin with the characters up to
the asterisk in the name given.  For example, "disable mod*" will
disable modsd, modsw, modud and moduw.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>