simple_random: Test unaligned loads and stores
Add an explicit 'align' field for the load/store instructions so we
can test unaligned loads and stores.  Most loads and stores have this
set to 1; the exceptions are DS-form loads and stores, which have it
set to 4, and the larx/stcx instructions, where it is the same as the
access size.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>