Add store-conditional instructions
This adds the st*cx. instructions into the mix.  This adds some
complications in that with the posix backend it is possible to lose
the reservation because of a context switch during the execution
of the instruction sequence.  That makes the result nondeterministic.
To cope with that, the posix backend now runs the instruction
sequence 5 times and takes the result which comes up most often.
This means we need to clear the memory before each run, so the
memory clearing has been pushed down into the backends.  We also
need to clear any pending reservation before each run.

The lwarx instruction was missing from the mix, so it gets added
as well.

This also prints out the contents of memory when registers=1.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>