Add sr_bisect script
sr_bisect is an expect script that connects to two Microwatt instances
that are running simple_random and compares their outputs for a range
of seed values.  When a difference is found, it then uses a bisection
search on the number of instructions to find the first instruction
where the results differ.  It then prints the registers from both
instances and the last (up to) 10 instructions generated.

The instances can be invoked by running a command (for instance,
"ssh arty ./simple_random") or connecting to a serial port.  There are
options to set the starting seed, the number of seed values to test,
and the maximum number of instructions (these default to 0, 1000 and
8000 respectively).  If the simple_random instances print FP/VEC/VSX
register contents, then give the '-v' flag to sr_bisect so it knows to
expect them.

An example invocation is:

$ sr_bisect 'ssh arty ./simple_random' /dev/ttyUSB0 -s 1234

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>